refactor: separate main functionality into distinct modules for clarity
